+++ src/App.js
@@ -3,8 +3,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Nav from 'react-bootstrap/Nav';
 import UserList from './pages/userList/userList';
-import {BasketPieChart, Pies, TotalPieChart} from './pages/basket/basket';
-import MealChart from './pages/meal/mealChart'; 
+import {MealPieChart, Pies} from './pages/gptCount/gptCount';
 import DashBoard from './pages/dashBoard/dashBoard';
 
 function Navi() {
@@ -17,7 +16,7 @@ function Navi() {
         <Nav.Link href="/adminPage/userList" eventKey="link-2">회원관리</Nav.Link>
       </Nav.Item>
       <Nav.Item>
-        <Nav.Link href="/adminPage/basket" eventKey="link-2">재료추출</Nav.Link>
+        <Nav.Link href="/adminPage/basket" eventKey="link-2">GPT Chart</Nav.Link>
       </Nav.Item>
       <Nav.Item>
         <Nav.Link href="/adminPage/shop" eventKey="link-3">공동구매</Nav.Link>
@@ -46,7 +45,7 @@ function App() {
           <Route path='/adminPage/main' element={<DashBoard/>}></Route>
           <Route path='/adminPage/basket' element={<Pies/>}></Route>
           <Route path="/adminPage/userList" element={<UserList/>}/>
-          <Route path="/adminPage/mealChart" element={<MealChart/>}/>
+          <Route path="/adminPage/mealChart" element={<MealPieChart/>}/>
         </Routes>
       </BrowserRouter>
     </div>
