--- src/pages/meal/mealChart.js
@@ -1,179 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { ResponsivePie } from '@nivo/pie';
-import axios from 'axios';
-
-const MealChart = () => {
-
-    const [mealCount, setMealCount] = useState([]);
-
-    useEffect(() => {
-        axios.post('/data/selectMealCount')
-            .then((resp) => {
-                console.log(resp);
-
-                let tmp = [
-                    {
-                      "id": "mealSuccessTotal",
-                      "label": "Success",
-                      "value":  resp.data.mealSuccessTotal,
-                      "color": "hsl(163, 70%, 50%)"
-                    },
-                    {
-                      "id": "mealFailTotal",
-                      "label": "Fail",
-                      "value": resp.data.mealFailTotal,
-                      "color": "hsl(328, 70%, 50%)"
-                    }
-                ];
-
-                setMealCount(tmp);
-            });
-
-    }, [])
-
-    const handle = {
-        padClick: (mealCount) => {
-            console.log(mealCount);
-        },
-
-        legendClick: (mealCount) => {
-            console.log(mealCount);
-        },
-    };
-
-    return (
-        // chart height이 100%이기 때문이 chart를 덮는 마크업 요소에 height 설정
-        <div style={{ width: '800px', height: '500px', margin: '0 auto' }}>
-            <ResponsivePie
-                data={mealCount}
-                margin={{ top: 40, right: 80, bottom: 80, left: 80 }}
-                innerRadius={0.5}
-                padAngle={0.7}
-                cornerRadius={3}
-                activeOuterRadiusOffset={8}
-                colors={{ scheme: 'paired' }}
-                borderWidth={1}
-                borderColor={{
-                    from: 'color',
-                    modifiers: [
-                        [
-                            'darker',
-                            0.2
-                        ]
-                    ]
-                }}
-                arcLinkLabelsSkipAngle={10}
-                arcLinkLabelsTextColor="#333333"
-                arcLinkLabelsThickness={2}
-                arcLinkLabelsColor={{ from: 'color' }}
-                arcLabelsSkipAngle={10}
-                arcLabelsTextColor={{
-                    from: 'color',
-                    modifiers: [
-                        [
-                            'darker',
-                            2
-                        ]
-                    ]
-                }}
-                defs={[
-                    {
-                        id: 'dots',
-                        type: 'patternDots',
-                        background: 'inherit',
-                        color: 'rgba(255, 255, 255, 0.3)',
-                        size: 4,
-                        padding: 1,
-                        stagger: true
-                    },
-                    {
-                        id: 'lines',
-                        type: 'patternLines',
-                        background: 'inherit',
-                        color: 'rgba(255, 255, 255, 0.3)',
-                        rotation: -45,
-                        lineWidth: 6,
-                        spacing: 10
-                    }
-                ]}
-                fill={[
-                    {
-                        match: {
-                            id: 'ruby'
-                        },
-                        id: 'dots'
-                    },
-                    {
-                        match: {
-                            id: 'c'
-                        },
-                        id: 'dots'
-                    },
-                    {
-                        match: {
-                            id: 'go'
-                        },
-                        id: 'dots'
-                    },
-                    {
-                        match: {
-                            id: 'python'
-                        },
-                        id: 'dots'
-                    },
-                    {
-                        match: {
-                            id: 'scala'
-                        },
-                        id: 'lines'
-                    },
-                    {
-                        match: {
-                            id: 'lisp'
-                        },
-                        id: 'lines'
-                    },
-                    {
-                        match: {
-                            id: 'elixir'
-                        },
-                        id: 'lines'
-                    },
-                    {
-                        match: {
-                            id: 'javascript'
-                        },
-                        id: 'lines'
-                    }
-                ]}
-                legends={[
-                    {
-                        anchor: 'bottom',
-                        direction: 'row',
-                        justify: false,
-                        translateX: 0,
-                        translateY: 56,
-                        itemsSpacing: 0,
-                        itemWidth: 100,
-                        itemHeight: 18,
-                        itemTextColor: '#999',
-                        itemDirection: 'left-to-right',
-                        itemOpacity: 1,
-                        symbolSize: 18,
-                        symbolShape: 'circle',
-                        effects: [
-                            {
-                                on: 'hover',
-                                style: {
-                                    itemTextColor: '#000'
-                                }
-                            }
-                        ]
-                    }
-                ]}
-            />
-        </div>
-    );
-};
-
-export default MealChart;
